# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: SEIP API
  version: 2.4.0
  summary: School Evidence Intelligence Platform — core API (evidence workflow + PA evaluation + reports/AI + PDF)
  description: |
    Contract v2.4 (CCR-009 cleanup) — list evidence carries aggregate `scan_status`;
    assignment detail carries `framework_version_id`. v2.3 (CCR-008) session
    lifecycle; v2.2 PDF; core remains evidence, PA evaluation, reports, AI suggest.

    **Business rules that bind every implementation**
    - Tenancy: every resource is school-scoped; cross-school access returns RES-001 (indistinguishable from not-found).
    - Taxonomy (frameworks/domains/indicators/weights) is versioned seed data — never hard-coded.
    - Scoring grain is per-evaluator (3 committee members); overall pass requires **every** evaluator ≥ 70% and the workload gate met.
    - Rounds are configurable 1..n per cycle; no endpoint assumes exactly 2.
    - Reports cite evidence by reference only (ReportSectionRef) — never embed file bytes.
    - PDF is generated from structured payload (not a pixel-perfect ก.ค.ศ. plate); suitable for draft/print review.
    - AI mapping suggestions always land as mapping_source=ai_suggested + status=suggested; human confirm required.
    - Error bodies follow error-codes.yaml; codes are stable API surface.

    **Deferred** — see `x-deferred`.
  license:
    name: Proprietary — SEIP project internal
    url: https://github.com/sooksun/aischool
  contact:
...
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /rounds/{roundId}/assignments:
    get:
      operationId: listAssignments
      summary: Evaluation assignments in a round
      tags: [scoring]
      parameters:
        - $ref: '#/components/parameters/RoundId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paged assignments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssignmentPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createAssignment
      summary: Create an assignment with its 3-member committee
      description: Committee must be exactly 3 (SCORE-001) — chair + 2 members, per ว9/ว10.
      tags: [scoring]
      parameters:
        - $ref: '#/components/parameters/RoundId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /assignments/{assignmentId}:
    get:
      operationId: getAssignment
      summary: Assignment detail (committee, workload gate, scoring progress)
      tags: [scoring]
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Assignment detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssignmentDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /assignments/{assignmentId}/my-scores:
    put:
      operationId: submitMyScores
      summary: Submit or replace the calling evaluator's full score set (idempotent upsert)
      description: |
        Per-evaluator grain (ADR-0003): each of the 3 committee members submits their own
        complete set — all scored standard indicators + challenge items (SCORE-005) with
        rubric levels 1..4 (SCORE-003). Requires committee membership (PERM-003),
        a declared workload gate (SCORE-004), and an open/scoring round (SCORE-002).
        Rollup (part1/part2/total %, ≥70% flag) is computed server-side per evaluator.
      tags: [scoring]
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScoreSubmission' }
      responses:
        '200':
          description: Stored; per-evaluator rollup returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvaluatorResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /assignments/{assignmentId}/results:
    get:
      operationId: getAssignmentResults
      summary: Per-evaluator rollups and overall pass verdict
      description: Overall pass = workload gate met AND every evaluator's total ≥ 70%.
      tags: [scoring]
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: Results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssignmentResults' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RoundId:
      name: roundId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: AUTH-* — missing/expired/invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: PERM-* — role or tenancy denial
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: RES-001 — not found or cross-tenant (indistinguishable by design)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnprocessableEntity:
      description: VAL/UPL/MAP/SCORE/CYCLE-* — semantically invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: RES-002/MAP-001/UPL-004/CYCLE-002/SCORE-002 — state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable code from error-codes.yaml }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            required: [field, issue]
            properties:
              field: { type: string }
              issue: { type: string }
        request_id: { type: string, format: uuid }
...
    RubricLevel:
      type: integer
      minimum: 1
      maximum: 4
...
    Assignment:
      type: object
      required: [id, round_id, evaluatee_personnel_id, status, committee]
      properties:
        id: { type: string, format: uuid }
        round_id: { type: string, format: uuid }
        evaluatee_personnel_id: { type: string, format: uuid }
        agreement_id: { type: [string, 'null'], format: uuid }
        status: { type: string, enum: [pending, in_progress, completed, void] }
        committee:
          type: array
          items: { $ref: '#/components/schemas/CommitteeMember' }
    CommitteeMember:
      type: object
      required: [evaluator_user_id, committee_role, seat_number]
      properties:
        evaluator_user_id: { type: string, format: uuid }
        committee_role: { type: string, enum: [chair, member] }
        seat_number: { type: integer, minimum: 1, maximum: 3 }
    AssignmentCreate:
      type: object
      required: [evaluatee_personnel_id, committee]
      properties:
        evaluatee_personnel_id: { type: string, format: uuid }
        agreement_id: { type: [string, 'null'], format: uuid }
        committee:
          type: array
          minItems: 3
          maxItems: 3
          description: Exactly 3 per ว9/ว10 (SCORE-001) — one chair, two members
          items: { $ref: '#/components/schemas/CommitteeMember' }
    AssignmentDetail:
      allOf:
        - $ref: '#/components/schemas/Assignment'
        - type: object
          required: [workload_gate_declared, my_submission_state, framework_version_id]
          properties:
            workload_gate_declared: { type: boolean }
            workload_met: { type: [boolean, 'null'] }
            my_submission_state:
              type: string
              enum: [not_member, not_submitted, submitted]
              description: Scoring progress of the calling evaluator
            framework_version_id:
              type: string
              format: uuid
              description: |
                Framework pinned by the assignment's cycle (CCR-009). Clients use this
                to load taxonomy for scoring — no graph walk over cycles/rounds.
    ScoreSubmission:
      type: object
      required: [indicator_scores]
      properties:
        workload_met:
          type: [boolean, 'null']
          description: Chair may declare the workload gate in the same call (SCORE-004)
        indicator_scores:
          type: array
          minItems: 1
          description: Complete set — every scored indicator (standard + challenge) of the framework (SCORE-005)
          items:
            type: object
            required: [indicator_id, rubric_level]
            properties:
              indicator_id: { type: string, format: uuid }
              rubric_level: { $ref: '#/components/schemas/RubricLevel' }
              comment: { type: [string, 'null'], maxLength: 2000 }
    EvaluatorResult:
      type: object
      required: [assignment_id, evaluator_user_id, part1_percent, part2_percent, total_percent, passed_individual_threshold]
      properties:
        assignment_id: { type: string, format: uuid }
        evaluator_user_id: { type: string, format: uuid }
        part1_percent: { type: number, description: 'Weighted per framework data (e.g. /60)' }
        part2_percent: { type: number }
        total_percent: { type: number }
        passed_individual_threshold: { type: boolean, description: total_percent >= 70 }
        computed_at: { type: string, format: date-time }
    AssignmentResults:
      type: object
      required: [assignment_id, workload_gate_met, evaluator_results, overall_pass, complete]
      properties:
        assignment_id: { type: string, format: uuid }
        workload_gate_met: { type: [boolean, 'null'] }
        evaluator_results:
          type: array
          items: { $ref: '#/components/schemas/EvaluatorResult' }
        complete: { type: boolean, description: All 3 evaluators submitted }
        overall_pass:
          type: [boolean, 'null']
          description: Null until complete; true iff workload gate met AND every evaluator >= 70%
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    AssignmentPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Assignment' }
        meta: { $ref: '#/components/schemas/PageMeta' }
...